refactor(duoplane): split up components
